Fix issues with SEO and localisation
We use a optional routing argument for locales which resulted in random
pages on depth 1 all ending up in the '/' route. This resulted in
robots.txt yielding a result which it should not have. In turn that
dropped our SEO ranking.

We also dynamically changanged language due to browser defaults, this is
not the recommended way to do it since. It is better to have different
routes for different languages according to googles SEO guied. Thus we
removed the '/sv/*' route and made the none localized route swedish.

We also added tests to validate that the behaviour is as expected, we
redirected any route from '/sv/*' -> '/*'.

Our URL structure was english, since the site will primarily be used in
Sweden we change the structure to be swedish, it will allow breadcrumbs
to look better in search engines.

closes #18, #16, #14