Move i18n implementation to seperate package
As we migrate to skeleton ui, we want to be able to reuse the i18n code.

We don't need translations to be reactive. We can load translation files on the
server, send them over and create our helper functions for interacting
with the files. This gives us better performance and simpler code.

The end result is a better performance score since the bundle size is
much smaller.