Setup directus docker-compose for local development
Before pushing to production we want to be able to test run changes
locally. To do this we have added a docker-compose file that builds our
directus container and runs it. We have then modified our sveltekit
project to use the local directus instance when running in development
and our public directus container in production.