Create vite plugin for generating static sitemap
Before the sitemap used the graphql client to fetch data from
cms.curarehab.se in theory this was a security issue since it would
allow us to be ddossed by pinging it to much depleating our database
credits.

By creating the sitemap.xml as a static file at build time we no longer
have this issue, as a bonus we also add lastmod to a couple of pages.